# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.ChapterMapsResponse:
    properties:
      chapter:
        $ref: '#/definitions/models.Chapter'
      maps:
        items:
          $ref: '#/definitions/models.MapSelect'
        type: array
    type: object
  handlers.ChaptersResponse:
    properties:
      chapters:
        items:
          $ref: '#/definitions/models.Chapter'
        type: array
      game:
        $ref: '#/definitions/models.Game'
    type: object
...
  handlers.CreateMapSummaryRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      record_date:
        type: string
      score_count:
        type: integer
      showcase:
        type: string
      user_name:
        type: string
    required:
    - category_id
    - description
    - record_date
    - score_count
    - user_name
    type: object
  handlers.DeleteMapSummaryRequest:
    properties:
      route_id:
        type: integer
    required:
    - route_id
    type: object
...
  handlers.EditMapSummaryRequest:
    properties:
      description:
        type: string
      record_date:
        type: string
      route_id:
        type: integer
      score_count:
        type: integer
      showcase:
        type: string
      user_name:
        type: string
    required:
    - description
    - record_date
    - route_id
    - score_count
    - user_name
    type: object
...
  handlers.MapLeaderboardsResponse:
    properties:
      map:
        $ref: '#/definitions/models.Map'
      pagination:
        $ref: '#/definitions/models.Pagination'
      records: {}
    type: object
...
  handlers.MapSummaryResponse:
    properties:
      map:
        $ref: '#/definitions/models.Map'
      summary:
        $ref: '#/definitions/models.MapSummary'
    type: object
...
  models.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.CategoryPortal:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      portal_count:
        type: integer
    type: object
  models.Chapter:
    properties:
      id:
        type: integer
      image:
        type: string
      is_disabled:
        type: boolean
      name:
        type: string
    type: object
  models.Game:
    properties:
      category_portals:
        items:
          $ref: '#/definitions/models.CategoryPortal'
        type: array
      id:
        type: integer
      image:
        type: string
      is_coop:
        type: boolean
      name:
        type: string
    type: object
...
  models.Map:
    properties:
      chapter_name:
        type: string
      game_name:
        type: string
      id:
        type: integer
      image:
        type: string
      is_coop:
        type: boolean
      is_disabled:
        type: boolean
      map_name:
        type: string
    type: object
  models.MapHistory:
    properties:
      date:
        type: string
      runner_name:
        type: string
      score_count:
        type: integer
    type: object
  models.MapRoute:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      completion_count:
        type: integer
      description:
        type: string
      history:
        $ref: '#/definitions/models.MapHistory'
      rating:
        type: number
      route_id:
        type: integer
      showcase:
        type: string
    type: object
  models.MapSelect:
    properties:
      category_portals:
        items:
          $ref: '#/definitions/models.CategoryPortal'
        type: array
      difficulty:
        type: integer
      id:
        type: integer
      image:
        type: string
      is_disabled:
        type: boolean
      name:
        type: string
    type: object
...
  models.MapSummary:
    properties:
      routes:
        items:
          $ref: '#/definitions/models.MapRoute'
        type: array
    type: object
  models.Pagination:
    properties:
      current_page:
        type: integer
      page_size:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
  models.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: lp.ardapektezol.com
info:
  contact: {}
  description: Backend API endpoints for the Least Portals Database.
  license:
    name: GNU Affero General Public License, Version 3
    url: https://www.gnu.org/licenses/agpl-3.0.html
  title: Least Portals Database API
  version: "1.0"
paths:
  /chapters/{chapterid}:
    get:
      description: Get maps from the specified chapter id.
      parameters:
      - description: Chapter ID
        in: path
        name: chapterid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.ChapterMapsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - games & chapters
...
  /games:
    get:
      description: Get games from the leaderboards.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Game'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - games & chapters
  /games/{gameid}:
    get:
      description: Get chapters from the specified game id.
      parameters:
      - description: Game ID
        in: path
        name: gameid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.ChaptersResponse'
              type: object
      tags:
      - games & chapters
  /games/{gameid}/maps:
    get:
      description: Get maps from the specified game id.
      parameters:
      - description: Game ID
        in: path
        name: gameid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.ChaptersResponse'
              type: object
      tags:
      - games & chapters
...
  /maps/{mapid}/leaderboards:
    get:
      description: Get map leaderboards with specified id.
      parameters:
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      - description: 'Page Number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of Records Per Page (default: 20)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.MapLeaderboardsResponse'
              type: object
      tags:
      - maps / leaderboards
...
  /maps/{mapid}/summary:
    delete:
      description: Delete map summary with specified map id.
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.DeleteMapSummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.DeleteMapSummaryRequest'
              type: object
      tags:
      - maps / summary
    get:
      description: Get map summary with specified id.
      parameters:
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.MapSummaryResponse'
              type: object
      tags:
      - maps / summary
    post:
      description: Create map summary with specified map id.
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateMapSummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.CreateMapSummaryRequest'
              type: object
      tags:
      - maps / summary
    put:
      description: Edit map summary with specified map id.
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.EditMapSummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.EditMapSummaryRequest'
              type: object
      tags:
      - maps / summary
...
swagger: "2.0"